Add enableConnectInNewWindow preference controlling connect button
